Fix java coding standard and improve code quality based on peer review
