+++ src/main/java/duke/util/Parser.java
@@ -1,6 +1,13 @@
 package duke.util;
 
-import duke.command.*;
+import duke.command.AddCommand;
+import duke.command.Command;
+import duke.command.DeleteCommand;
+import duke.command.DoneCommand;
+import duke.command.ExitCommand;
+import duke.command.FindCommand;
+import duke.command.ListCommand;
+import duke.exceptions.UnclearInstructionException;
 import duke.exceptions.UnknownException;
 
 /**
@@ -34,5 +41,59 @@ public class Parser {
         } else {
             throw new UnknownException();
         }
+    }
+
+    /**
+     * Returns task description of a full command.
+     *
+     * @param text  Full command given by user.
+     * @return Task description.
+     * @throws UnclearInstructionException  If the given command cannot be properly extracted or has empty description.
+     */
+    public static String extractTaskDescription(String text) throws UnclearInstructionException {
+        String[] contents = text.split(" ", 3);
+        String task_type = contents[0];
+        String description = "";
+
+        if (contents.length != 3) {
+            throw new UnclearInstructionException(
+                    "OOPS!!! The description of " + task_type + " cannot be extracted properly.");
+        }
+
+        int istart = text.indexOf(" ");
+        int iend = text.indexOf("/");
+
+        description = text.substring(istart + 1, iend - 1);
+
+        if (description.equals("")) {
+            throw new UnclearInstructionException(
+                    "OOPS!!! The description of " + task_type + " cannot be empty.");
+        }
+        return description;
+    }
+
+    /**
+     * Returns event time of a full command.
+     *
+     * @param text  Full command given by user.
+     * @return Task time.
+     * @throws UnclearInstructionException If the given command cannot be properly extracted or has empty time.
+     */
+    public static String extractTaskTime(String text) throws UnclearInstructionException {
+        String[] contents = text.split(" ", 3);
+        String task_type = contents[0];
+        if (contents.length != 3) {
+            throw new UnclearInstructionException(
+                    "OOPS!!! The description of " + task_type + " cannot be extracted properly.");
+        }
+
+        int istart = text.indexOf(" ");
+        int iend = text.indexOf("/");
+        String time = text.substring(iend + 4);
+
+        if (time.equals("")) {
+            throw new UnclearInstructionException("OOPS!!! The time of " + task_type + " cannot be empty.");
+        }
+        return time;
     }
 }
